Added more functions

+++ commands/routes/responder.js
@@ -18,9 +18,14 @@ var state = {}
 
 router.post('/upload', upload.single('file'), (req, res) => {
     let fileName = req.file.originalname
+    if (fileName === undefined) {
+        res.sendStatus(400);
+        console.log("No file to upload");
+        return
+    }
     //delete all other files
     fs.readdirSync('uploads/').forEach(file => {
-        if (file !== fileName) {
+        if (file !== fileName && fs.existsSync('uploads/' + file)) {
             fs.unlinkSync('uploads/' + file);
         }
     });
@@ -28,7 +33,7 @@ router.post('/upload', upload.single('file'), (req, res) => {
 });
 
 router.get('/download', (req, res) => {
-    let fileName = fs.readdirSync('uploads/')[0];
+    let fileName = req.query.file || fs.readdirSync('uploads/')[0];
     if (fileName === undefined) {
         res.sendStatus(404);
         console.log("No file to download");
@@ -37,9 +42,25 @@ router.get('/download', (req, res) => {
     res.download(`uploads/${fileName}`);
 });
 
-router.get('/ls', (req, res) => {
-    let files = fs.readdirSync('uploads/');
-    res.send(files);
+router.get('/do', (req, res) => {
+    let command = req.query.command;
+    if(command === "list"){
+        let files = fs.readdirSync('uploads/');
+        res.send(files);
+    } else if (command === "delete") {
+        let fileName = req.query.file;
+        if (fileName === undefined) {
+            res.sendStatus(404);
+            console.log("No file to delete");
+            return
+        }
+        if (fs.existsSync('uploads/' + fileName)) {
+            fs.unlinkSync('uploads/' + fileName);
+            res.status(200).send("Deleted " + fileName);
+        } else {
+            res.sendStatus(404);
+        }
+    }
 });
 
 module.exports = router;
